[SWITCHYARD-340] Refactored split package, fixed properties and feature dependencies
